рефакторинг utils
