+++ src/mock/comments.js
@@ -1,5 +1,5 @@
 import dayjs from 'dayjs';
-import { getRandomInteger } from './utils-for-mock.js';
+import { getRandomInteger } from '../utils.js/utils-for-mock.js';
 
 const generateText = () => {
   const texts = [
