--- src/view/utils-for-view.js
@@ -1,2 +1,0 @@
-export const changeDateFormatToYear = (date) => date.format('YYYY');
-export const changeDateFormatToFull = (date) => date.format('DD MMMM YYYY');
